Day-9/hackathon/regulation-aware-site-auditor/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4>=4.12.3,<5.0.0
google-generativeai<0.8.0, >=0.7.0
python-dotenv==1.0.0
validators==0.22.0
pydantic==2.9.2
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6
crewai==0.41.1
crewai-tools==0.4.26
chromadb >=0.4.22, <0.5.0
langchain-google-genai==1.0.10
langchain-community==0.2.16
selenium>=4.18.1,<5.0.0 
webdriver-manager==4.0.1 